Implement basic permissions to delete programs

+++ app/Http/Controllers/ProgramController.php
@@ -92,7 +92,7 @@ class ProgramController extends Controller
      */
     public function destroy(Program $program)
     {
-        // $this->authorize('destroy');
+        $this->authorize('destroy');
 
         $program->delete();
